Fix HMT on redux
